Separate bisection from base class
