+++ Anima/math-tools/optimizers/animaBisectionRootFindingAlgorithm.h
@@ -1,0 +1,17 @@
+#pragma once
+#include "AnimaOptimizersExport.h"
+
+#include <animaBaseRootFindingAlgorithm.h>
+
+namespace anima
+{
+
+class ANIMAOPTIMIZERS_EXPORT BisectionRootFindingAlgorithm : public BaseRootFindingAlgorithm
+{
+public:
+    using Superclass = BaseRootFindingAlgorithm;
+    using ParametersType = Superclass::ParametersType;
+    double Optimize() ITK_OVERRIDE;
+};
+
+} // end namespace anima
